+++ vue/src/index.ts
@@ -11,5 +11,6 @@ export { default as ImportScheduler } from './ImportScheduler/ImportScheduler.vu
 export { default as ImportSchedulerGA4 } from './ImportScheduler/ImportSchedulerGA4.vue';
 export { default as ImportSelector } from './ImportScheduler/ImportSelector.vue';
 export { default as ImportStatus } from './ImportStatus/ImportStatus.vue';
+export { default as ConfigureConnection } from './Configure/ConfigureConnection.vue';
 export { default as ClientConfig } from './ClientConfig/ClientConfig.vue';
 export { default as AdminPage } from './AdminPage/AdminPage.vue';